change way sending SCP state works after connection
The previous flow was:
- after connection node sends SCP envelopes for last 3 slot indexes to
just connected node

The target flow of sending SCP state is:
- after connection node request SCP state since its LCL
- node send requested state - all envelopes with slot index bigger or
equal to expected

Current flow is combination of these two. First one is kept to let nodes
that are not yet upgraded get at least some of needed SCP state. After
all nodes are upgraded first part can be removed.

This code ensures that only required envelopes will be sent.

This commit changes way how GET_SCP_STATE message works, but it was not
used before, so it does not break anything.

Signed-off-by: Rafał Malinowski <rafal.przemyslaw.malinowski@gmail.com>